Don't use raise_for_status by default
The use of `raise_for_status` in our client session was preventing us from returning detailed information about Salesforce errors.

For example, when we got authorization errors previously, we were raising an `aiohttp.ClientResponseError` with `status=401` and `message=Unauthorized`.

The new behavior is that we raise a `SalesforceRestApiError` with `message=Session expired or invalid` and `error_code=INVALID_SESSION_ID`.